requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
requests==2.31.0
beautifulsoup4==4.12.2
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.3.0  # Atualizado para versão compatível com Python 3.12+
spacy==3.7.2
scikit-learn==1.4.0
numpy==1.26.0
python-dateutil==2.8.2
pt_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.7.0/pt_core_news_lg-3.7.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.7.0/pt_core_news_lg-3.7.0.tar.gz#egg=pt_core_news_lg
https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.7.0/pt_core_news_lg-3.7.0.tar.gz#egg=pt_core_news_lg
https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.7.0/pt_core_news_lg-3.7.0.tar.gz#egg=pt_core_news_lg
